Moved SDK file to src directory
