+++ src/CampfireSDK.js
